+++ convertkit/includes/class-wp-convertkit.php
@@ -96,7 +96,6 @@ class WP_ConvertKit {
 		$this->classes['admin_setup_wizard_landing_page']     = new ConvertKit_Admin_Setup_Wizard_Landing_Page();
 		$this->classes['admin_setup_wizard_plugin']           = new ConvertKit_Admin_Setup_Wizard_Plugin();
 		$this->classes['admin_setup_wizard_restrict_content'] = new ConvertKit_Admin_Setup_Wizard_Restrict_Content();
-		$this->classes['admin_tinymce']                       = new ConvertKit_Admin_TinyMCE();
 
 		/**
 		 * Initialize integration classes for the WordPress Administration interface.
@@ -183,6 +182,7 @@ class WP_ConvertKit {
 	private function initialize_global() {
 
 		$this->classes['admin_notices']                               = new ConvertKit_Admin_Notices();
+		$this->classes['admin_tinymce']                               = new ConvertKit_Admin_TinyMCE();
 		$this->classes['admin_refresh_resources']                     = new ConvertKit_Admin_Refresh_Resources();
 		$this->classes['blocks_convertkit_broadcasts']                = new ConvertKit_Block_Broadcasts();
 		$this->classes['blocks_convertkit_content']                   = new ConvertKit_Block_Content();